feat: add TMDB link button for external movie view
- created TMDBLinkButtonView to open movie page on TMDB
- linked to movie.tmdbURL using SwiftUI's Link
- added button to MovieDetailView for easy access
